Introduce the template system
Add template system, to increase flexibility and maintainability of the
code.

* Optimization note:

The template system is a bottleneck for the TTY plots, since it adds
about 200 ms of execution time. This could be undesirable in realtime
applications. One (rather inelegant) alternative is to avoid templates
in the plotext plots, and just implement them directly in the main
script.

However, sice the delay happens before the flushing of the data to
stdout, this overhead doesn't result in stutter or flashing of the
terminal.